게시글 삭제 후 406 Error 해결

+++ src/app/(afterLogin)/[userId]/posts/[postId]/_lib/deletePost.ts
@@ -14,8 +14,7 @@ export async function deletePost(postId: number, userId: string | undefined) {
     .from("posts")
     .delete()
     .eq("id", postId)
-    // .eq("user_id", userId);
-    .maybeSingle();
+    .eq("user_id", userId);
 
   if (error) {
     console.error("Error deleting post:", error.message);
